Tools/auterion: enable reverting to release firmware
using the command:
./Tools/auterion/upload_skynode.sh --revert
possibly with a --wifi at the end if not over USB

+++ Tools/auterion/remote_update_fmu.sh
@@ -1,14 +1,14 @@
 #!/bin/bash
 # Flash PX4 to a device running AuterionOS in the local network
 if [ "$1" == "-h" ] || [ "$1" == "--help" ] || [ $# -lt 2 ]; then
-	echo "Usage: $0 -f <firmware.px4|.elf> [-c <configuration_dir>] -d <IP/Device> [-u <user>] [-p <ssh_port>]"
+	echo "Usage: $0 -f <firmware.px4|.elf> [-c <configuration_dir>] -d <IP/Device> [-u <user>] [-p <ssh_port>] [--revert]"
 	exit 1
 fi
 
 ssh_port=22
 ssh_user=root
 
-while getopts ":f:c:d:p:u:" opt; do
+while getopts ":f:c:d:p:u:r" opt; do
 	case ${opt} in
 		f )
 			if [ -n "$OPTARG" ]; then
@@ -50,53 +50,66 @@ while getopts ":f:c:d:p:u:" opt; do
 				exit 1
 			fi
 			;;
+		r )
+			revert=true
+			;;
 	esac
 done
-
-target_dir=/shared_container_dir/fmu
-
-# create update-dev.tar
-target_file_name="update-dev.tar"
-tmp_dir="$(mktemp -d)"
-config_path=""
-firmware_path=""
-
-if [ -d "$config_dir" ]; then
-	cp -r "$config_dir" "$tmp_dir/config"
-	config_path=config
-fi
-
-if [ -f "$firmware_file" ]; then
-	extension="${firmware_file##*.}"
-	cp "$firmware_file" "$tmp_dir/firmware.$extension"
-	if [ "$extension" == "elf" ]; then
-		# ensure the file is stripped to reduce file size
-		arm-none-eabi-strip "$tmp_dir/firmware.$extension"
-	fi
-	firmware_path="firmware.$extension"
-fi
 
 if [ -z "$device" ]; then
 	echo "Error: missing device"
 	exit 1
 fi
 
-pushd "$tmp_dir" &>/dev/null
+if [ "$revert" == true ]; then
+	# revert to the release version which was originally deployed
+	cmd="cp /persistent/shared_container_dir/fmu/update.tar /persistent/shared_container_dir/fmu/update-dev.tar"
+	ssh -t -p $ssh_port $ssh_user@$device "$cmd"
+else
+	# create custom update-dev.tar
+	tmp_dir="$(mktemp -d)"
+	config_path=""
+	firmware_path=""
 
-if [ -z $firmware_path ] && [ -z $config_path ]; then
-	exit 1
+	if [ -d "$config_dir" ]; then
+		cp -r "$config_dir" "$tmp_dir/config"
+		config_path=config
+	fi
+
+	if [ -f "$firmware_file" ]; then
+		extension="${firmware_file##*.}"
+		cp "$firmware_file" "$tmp_dir/firmware.$extension"
+		if [ "$extension" == "elf" ]; then
+			# ensure the file is stripped to reduce file size
+			arm-none-eabi-strip "$tmp_dir/firmware.$extension"
+		fi
+		firmware_path="firmware.$extension"
+	fi
+
+	pushd "$tmp_dir" &>/dev/null
+
+	if [ -z $firmware_path ] && [ -z $config_path ]; then
+		exit 1
+	fi
+
+	tar_name="tar"
+
+	if [ -x "$(command -v gtar)" ]; then
+		# check if gnu-tar is installed on macOS and use that instead
+		tar_name="gtar"
+	fi
+
+	target_dir=/shared_container_dir/fmu
+	target_file_name="update-dev.tar"
+
+	$tar_name -C "$tmp_dir" --sort=name --owner=root:0 --group=root:0 --mtime='2019-01-01 00:00:00' -cvf $target_file_name $firmware_path $config_path
+
+	# send it to the target to start flashing
+	scp -P $ssh_port "$target_file_name" $ssh_user@"$device":$target_dir
+	popd &>/dev/null
+	rm -rf "$tmp_dir"
 fi
 
-# check if gnu-tar is installed on macOS and use that instead
-tar_name="tar"
-if [ -x "$(command -v gtar)" ]; then
-	tar_name="gtar"
-fi
-$tar_name -C "$tmp_dir" --sort=name --owner=root:0 --group=root:0 --mtime='2019-01-01 00:00:00' -cvf $target_file_name $firmware_path $config_path
-scp -P $ssh_port "$target_file_name" $ssh_user@"$device":$target_dir
-popd &>/dev/null
-rm -rf "$tmp_dir"
-
-# grab status output
+# grab status output for flshing progress
 cmd="tail --follow=name $target_dir/update_status 2>/dev/null || true"
 ssh -t -p $ssh_port $ssh_user@$device "$cmd"
